Refactor TimezoneComboBox
